+++ mycore-base/src/main/java/org/mycore/frontend/servlets/MCRQRCodeServlet.java
@@ -92,7 +92,7 @@ public class MCRQRCodeServlet extends MCRContentServlet {
         MCRContent content = getPNGContent(url, size);
         content.setLastModified(0);
         if (!"HEAD".equals(req.getMethod())) {
-            MCRFrontendUtil.writeCacheHeaders(resp, CACHE_TIME, (long) 0, true);
+            MCRFrontendUtil.writeCacheHeaders(resp, CACHE_TIME, 0, true);
         }
         return content;
     }
